Fix delete and edit list bugs
1. For deleting the list, should not use componentDidUpdate.
In the reducer, I was not passing in the right value to action.
After passing in the correct id, the state changes after delete,
which led component to re-render automatically.

2. Edit list form modal didn't show the original name.
It was just showing the placeholder "Name".
Because the default list (with empty name) was passed.
But we should pass in the list to edit.
Problem was solved after I passed in the list id.